+++ CHANGELOG.md
@@ -8,5 +8,6 @@ All notable changes to the "Hydro Helper" extension will be documented in this f
 - Rename configuration keys
 - Add very simple walkthrough
 - Fix tree refresh
+- Refactor fetchData to use async/await for improved readability and error handling
 
 Want to see the previous log? [Click here](https://github.com/langningchen/hydro-helper/commits/main/CHANGELOG.md)